# docs/swagger.yaml
basePath: /subscriptions
definitions:
  domain.CreateSubscriptionInput:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        description: '@Schema(required=true)'
        example: 499
        type: integer
      service_name:
        description: '@Schema(required=true)'
        example: Netflix
        type: string
      start_date:
        description: '@Schema(required=true)'
        example: 07-2025
        type: string
      user_id:
        description: '@Schema(required=true)'
        example: 111e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  domain.Subscription:
    properties:
      created_at:
        example: "2025-01-01T12:00:00Z"
        type: string
      end_date:
        example: 12-2025
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      price:
        example: 499
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        example: "2025-01-01T12:00:00Z"
        type: string
      user_id:
        example: 111e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  domain.SumSubscriptionsFilter:
    properties:
      from:
        description: '@Schema(required=true)'
        example: 01-2025
        type: string
      service_name:
        description: '@Schema(required=true)'
        example: Netflix
        type: string
      to:
        description: '@Schema(required=true)'
        example: 12-2025
        type: string
      user_id:
        description: '@Schema(required=true)'
        example: 111e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  domain.UpdateSubscriptionInput:
    properties:
      end_date:
        example: 11-2025
        type: string
      price:
        example: 299
        type: integer
      service_name:
        example: Spotify
        type: string
      start_date:
        example: 08-2025
        type: string
    type: object
  lib.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  lib.SuccessResponse:
    properties:
      result:
        example: success
        type: string
    type: object
  sum.SuccessSumResponse:
    properties:
      amount:
        example: 1000
        type: integer
    type: object
info:
  contact:
    email: l.golofastov@mail.ru
    name: Lev Golofastov
  description: API for managing user subscriptions
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Subscriptions Manager API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Get all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create new subscription
      parameters:
      - description: Create subscription
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateSubscriptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: Get subscription
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Update subscription
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateSubscriptionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      consumes:
      - application/json
      description: Calculate total price of subscriptions
      parameters:
      - description: Sum filter
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SumSubscriptionsFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sum.SuccessSumResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      summary: Sum subscriptions prices
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"